mqtt_subscriber/pyproject.toml
[project]
name = "mqtt_subscriber"
version = "0.1.0"
authors = [
    { name = "Martin Pres", email = "martin.pres@prespol.cz" }
]
description = "Consumes data from MQTT topic and stores them to MongoDB and PostgreSQL databases."
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt==2.1.0",
    "psycopg[binary,pool]==3.2.3",
    "pymongo==4.10.1"
]

[project.scripts]
mqtt_subscriber = "subscriber.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false
